Cleanup of utils.js
Move funcs used in only one place out of utils
