+++ test/blocks/caas/mocks/utils.js
@@ -1,26 +1,12 @@
 import { stub } from 'sinon';
 
-export const isValidUuid = () => true;
-
-export const getHashConfig = stub();
-
 export const getConfig = () => ({});
-
-export const cloneObj = (obj) => JSON.parse(JSON.stringify(obj));
 
 export const loadStyle = stub();
 
 export const loadScript = stub();
 
 export const utf8ToB64 = (str) => window.btoa(unescape(encodeURIComponent(str)));
-
-export function updateObj(obj, defaultObj) {
-  const ds = cloneObj(defaultObj);
-  Object.keys(ds).forEach((key) => {
-    if (obj[key] === undefined) obj[key] = ds[key];
-  });
-  return obj;
-}
 
 export function createIntersectionObserver({ el, callback, once = true, options = {} }) {
   // fire immediately
