+++ test/utils/getUuid.test.js
@@ -2,7 +2,7 @@
 
 import { expect } from '@esm-bundle/chai';
 import getUuid from '../../../libs/utils/getUuid.js';
-import { isValidUuid } from '../../../libs/utils/utils.js';
+import { isValidUuid } from '../../../libs/blocks/caas-config/caas-config.js';
 
 const LONG_STRING = 'go milo'.repeat(2000);
 describe('Get a UUID from a string', () => {
